Update response condition
